// backend/src/app_document/dto/create-document.dto.ts
export interface IFields {
  court_name: string;
  court_address: string;
  plaintiff_fio_or_name: string;
  plaintiff_address: string;
  plaintiff_iin_bin: string;
  plaintiff_phone_email: string;
  defendant_fio_or_name: string;
  defendant_address: string;
  defendant_iin_bin: string;
  defendant_phone_email: string;
  date: string;
  signature: string;
  price_of_claim: string;
  case_number: string;
  representative: string;
  expert_type: string;
  expert_questions: string;
  expert_institution: string;
  hearing_date: string;
  protocol_date: string;
  document_id: number;
  situation: string;
}